gm marche

+++ src/Player.cpp
@@ -66,7 +66,7 @@ void Player::deleteItem(unsigned short int i){
     myItem[i] = NULL;
 }
 
-int Player::isDie(){
+short int Player::isDie(){
   for(unsigned int i = 0; i<nbPokemon; i++){
       if(myPokemon[i].getHp() > 0 ){
         //cout << "Pokemon : "<<i<<" alive" <<endl;
